Implement mySQL database unit tests using jest
Add connection variable to module.exports in the database index.js file so that it can be used in my test file.

Write 3 tests. The first tests for the correct number of restaurants in database. The second tests that the getRestaurantInfo method queries accurate information given a restaurant id. The last test tests to see if the getRestaurantInfo method properly handles an id that does not exist in the database.

Remove sum.test.js file which was initially setup to see if jest was working properly.